requirements.txt
discord.py==2.6.4
flask==2.3.5
psycopg2-binary==2.9.9
